Align with spec
